requirements.txt
annotated-types==0.7.0
colorama==0.4.6
et_xmlfile==2.0.0
iniconfig==2.1.0
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pluggy==1.6.0
pydantic==2.11.5
pydantic_core==2.33.2
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
pyyaml>=6.0
